Removed background image, refactored grievance form
